requirements.txt
fastapi
uvicorn
google-generativeai
python-multipart
jinja2
google-genai